app/requirements.txt
discord.py
python-dotenv
apscheduler
pytz
requests
opencensus-ext-azure
opencensus-ext-logging